Refactoring/alex

+++ moviepy/video/VideoClip.py
@@ -95,7 +95,7 @@ class VideoClip(Clip):
     """
 
     def __init__(
-        self, make_frame=None, is_mask=False, duration=None, has_constant_size=True
+            self, make_frame=None, is_mask=False, duration=None, has_constant_size=True
     ):
         super().__init__()
         self.mask = None
@@ -198,26 +198,26 @@ class VideoClip(Clip):
     @convert_masks_to_RGB
     @convert_path_to_string(["filename", "temp_audiofile", "temp_audiofile_path"])
     def write_videofile(
-        self,
-        filename,
-        fps=None,
-        codec=None,
-        bitrate=None,
-        audio=True,
-        audio_fps=44100,
-        preset="medium",
-        audio_nbytes=4,
-        audio_codec=None,
-        audio_bitrate=None,
-        audio_bufsize=2000,
-        temp_audiofile=None,
-        temp_audiofile_path="",
-        remove_temp=True,
-        write_logfile=False,
-        threads=None,
-        ffmpeg_params=None,
-        logger="bar",
-        pixel_format=None,
+            self,
+            filename,
+            fps=None,
+            codec=None,
+            bitrate=None,
+            audio=True,
+            audio_fps=44100,
+            preset="medium",
+            audio_nbytes=4,
+            audio_codec=None,
+            audio_bitrate=None,
+            audio_bufsize=2000,
+            temp_audiofile=None,
+            temp_audiofile_path="",
+            remove_temp=True,
+            write_logfile=False,
+            threads=None,
+            ffmpeg_params=None,
+            logger="bar",
+            pixel_format=None,
     ):
         """Write the clip to a videofile.
 
@@ -349,7 +349,7 @@ class VideoClip(Clip):
 
         audiofile = audio if isinstance(audio, str) else None
         make_audio = (
-            (audiofile is None) and (audio is True) and (self.audio is not None)
+                (audiofile is None) and (audio is True) and (self.audio is not None)
         )
 
         if make_audio and temp_audiofile:
@@ -392,16 +392,15 @@ class VideoClip(Clip):
             pixel_format=pixel_format,
         )
 
-        if remove_temp and make_audio:
-            if os.path.exists(audiofile):
-                os.remove(audiofile)
+        if remove_temp and make_audio and os.path.exists(audiofile):
+            os.remove(audiofile)
         logger(message="MoviePy - video ready %s" % filename)
 
     @requires_duration
     @use_clip_fps_by_default
     @convert_masks_to_RGB
     def write_images_sequence(
-        self, name_format, fps=None, with_mask=True, logger="bar"
+            self, name_format, fps=None, with_mask=True, logger="bar"
     ):
         """Writes the videoclip to a sequence of image files.
 
@@ -458,18 +457,18 @@ class VideoClip(Clip):
     @convert_masks_to_RGB
     @convert_path_to_string("filename")
     def write_gif(
-        self,
-        filename,
-        fps=None,
-        program="imageio",
-        opt="nq",
-        fuzz=1,
-        loop=0,
-        dispose=False,
-        colors=None,
-        tempfiles=False,
-        logger="bar",
-        pixel_format=None,
+            self,
+            filename,
+            fps=None,
+            program="imageio",
+            opt="nq",
+            fuzz=1,
+            loop=0,
+            dispose=False,
+            colors=None,
+            tempfiles=False,
+            logger="bar",
+            pixel_format=None,
     ):
         """Write the VideoClip to a GIF file.
 
@@ -593,11 +592,26 @@ class VideoClip(Clip):
         >>> new_clip = clip.subapply(lambda c:c.multiply_speed(0.5) , 3,6)
 
         """
-        left = None if (start_time == 0) else self.subclip(0, start_time)
-        center = self.subclip(start_time, end_time).fx(fx, **kwargs)
-        right = None if (end_time is None) else self.subclip(start_time=end_time)
+        # Left channel
+        if start_time == 0:
+            left = None
+        else:
+            left = self.subclip(0, start_time)
 
-        clips = [clip for clip in [left, center, right] if clip is not None]
+        # Center channel
+        center = self.subclip(start_time, end_time).fx(fx, **kwargs)
+
+        # Right channel
+        if end_time is None:
+            right = None
+        else:
+            right = self.subclip(start_time=end_time)
+
+        # Setup clips
+        clips = []
+        for clip in [left, center, right]:
+            if clip is not None:
+                clips.append(clip)
 
         # beurk, have to find other solution
         from moviepy.video.compositing.concatenate import concatenate_videoclips
@@ -764,9 +778,9 @@ class VideoClip(Clip):
             )
 
         if (
-            isinstance(self, ImageClip)
-            and (not hasattr(pos, "__call__"))
-            and ((self.mask is None) or isinstance(self.mask, ImageClip))
+                isinstance(self, ImageClip)
+                and (not hasattr(pos, "__call__"))
+                and ((self.mask is None) or isinstance(self.mask, ImageClip))
         ):
             new_result = result.to_ImageClip()
             if result.mask is not None:
@@ -1075,7 +1089,7 @@ class ImageClip(VideoClip):
     """
 
     def __init__(
-        self, img, is_mask=False, transparent=True, fromalpha=False, duration=None
+            self, img, is_mask=False, transparent=True, fromalpha=False, duration=None
     ):
         VideoClip.__init__(self, is_mask=is_mask, duration=duration)
 
@@ -1267,25 +1281,25 @@ class TextClip(ImageClip):
 
     @convert_path_to_string("filename")
     def __init__(
-        self,
-        text=None,
-        filename=None,
-        size=None,
-        color="black",
-        bg_color="transparent",
-        font_size=None,
-        font="Courier",
-        stroke_color=None,
-        stroke_width=1,
-        method="label",
-        kerning=None,
-        align="center",
-        interline=None,
-        tempfilename=None,
-        temptxt=None,
-        transparent=True,
-        remove_temp=True,
-        print_cmd=False,
+            self,
+            text=None,
+            filename=None,
+            size=None,
+            color="black",
+            bg_color="transparent",
+            font_size=None,
+            font="Courier",
+            stroke_color=None,
+            stroke_width=1,
+            method="label",
+            kerning=None,
+            align="center",
+            interline=None,
+            tempfilename=None,
+            temptxt=None,
+            transparent=True,
+            remove_temp=True,
+            print_cmd=False,
     ):
         if text is not None:
             if temptxt is None:
@@ -1428,7 +1442,7 @@ class BitmapClip(VideoClip):
 
     @convert_parameter_to_seconds(["duration"])
     def __init__(
-        self, bitmap_frames, *, fps=None, duration=None, color_dict=None, is_mask=False
+            self, bitmap_frames, *, fps=None, duration=None, color_dict=None, is_mask=False
     ):
         """Creates a VideoClip object from a bitmap representation. Primarily used
         in the test suite.
@@ -1508,22 +1522,28 @@ class BitmapClip(VideoClip):
         )
         self.fps = fps
 
-    def to_bitmap(self, color_dict=None):
+    def to_bitmap(self, letter_RGB_dict=None):
         """Returns a valid bitmap list that represents each frame of the clip.
         If `color_dict` is not specified, then it will use the same `color_dict`
         that was used to create the clip.
         """
-        color_dict = color_dict or self.color_dict
+        letter_RGB_dict = letter_RGB_dict or self.color_dict
 
         bitmap = []
         for frame in self.iter_frames():
             bitmap.append([])
             for line in frame:
                 bitmap[-1].append("")
-                for pixel in line:
-                    letter = list(color_dict.keys())[
-                        list(color_dict.values()).index(tuple(pixel))
-                    ]
-                    bitmap[-1][-1] += letter
-
+                for pixel_RGB in line:
+                    color_letter = get_letter(letter_RGB_dict, pixel_RGB)
+                    bitmap[-1][-1] += color_letter
         return bitmap
+
+
+def get_letter(color_dict, pixel_RGB):
+    pixel_RGB = tuple(pixel_RGB)
+    color_letters = color_dict.keys()
+    letter_RGBs = color_dict.values()
+
+    color_letter = list(color_letters)[list(letter_RGBs).index(pixel_RGB)]
+    return color_letter
